+++ src/org/sosy_lab/cpachecker/cpa/bdd/BDDVectorCExpressionVisitor.java
@@ -362,7 +362,7 @@ public class BDDVectorCExpressionVisitor
   }
 
   protected int getSize(CType pType) {
-    return machineModel.getSizeof(pType) * machineModel.getSizeofCharInBits();
+    return Math.toIntExact(machineModel.getSizeof(pType) * machineModel.getSizeofCharInBits());
   }
 
   /**
